+++ tests/presentation/components/cardList/view.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { Image, Text, View } from 'react-native';
 import { render } from '@testing-library/react-native';
 import { faker } from '@faker-js/faker';
+import { CardList } from '~/presentation/components';
 
 describe('Components: CardList', () => {
   test('should show title with success', () => {
@@ -69,34 +69,5 @@ const makeSut = (
       ratingComments={ratingComments}
       reviews={reviews}
     />,
-  );
-};
-
-type Props = {
-  imageUrl: string;
-  title: string;
-  location: string;
-  myDistanceOfLocal: string;
-  ratingComments: string;
-  reviews: string;
-};
-
-const CardList = ({
-  imageUrl,
-  title,
-  location,
-  myDistanceOfLocal,
-  ratingComments,
-  reviews,
-}: Props) => {
-  return (
-    <View>
-      <Image testID="image_of_place_id" source={{ uri: imageUrl }} />
-      <Text testID="title_id">{title}</Text>
-      <Text testID="location_id">{location}</Text>
-      <Text testID="distance_of_local_id">{myDistanceOfLocal}</Text>
-      <Text testID="rating_comments_id">{ratingComments}</Text>
-      <Text testID="reviews_id">{reviews}</Text>
-    </View>
   );
 };
